Move test classes to the test suite.
