+++ Native/Messages/Tests/Test.cs
@@ -1,4 +1,5 @@
 ﻿using Chopsticks.Messages;
+using Chopsticks.Messages.Handlers;
 using Chopsticks.Messages.Handlers.Multicast;
 using Chopsticks.Messages.Registration;
 
@@ -6,6 +7,56 @@ namespace Tests
 {
     public class Test
     {
+        public class Message
+        {
+            public string Value { get; set; }
+        }
+
+        public class MessageSender
+        {
+            private IMessageHandler<Message> _handler;
+
+            public MessageSender(IMessageHandler<Message> handler) =>
+                _handler = handler;
+            public HandlingPromise Send()
+            {
+                Console.WriteLine("Sending OnCommand");
+                var promise = _handler.Handle(new Message());
+                Console.WriteLine("Sent OnCommand");
+
+                return promise;
+            }
+
+            public async Task<HandlingResult> SendAsync()
+            {
+                Console.WriteLine("Sending OnCommand");
+                var result = await _handler.HandleAsync(new Message());
+                Console.WriteLine("Sent OnCommand");
+
+                return result;
+            }
+        }
+
+        public class MessageHandler : ISyncMessageHandler<Message>
+        {
+            void ISyncMessageHandler<Message>.Handle(Message command)
+            {
+                Console.WriteLine($"Received OnCommand, Sync, Value: {command.Value}.");
+            }
+        }
+
+        public class MessageAsyncHandler : ITaskMessageHandler<Message>
+        {
+            async Task ITaskMessageHandler<Message>.HandleAsync(
+                Message message, CancellationToken token)
+            {
+                Console.WriteLine($"Received OnCommand, Async, Value: {message.Value}.");
+                await Task.Delay(100); // Simulate async work.
+                Console.WriteLine($"Done handling OnCommand, Async, Value: {message.Value}.");
+            }
+        }
+
+
         [Test]
         public async Task Integration_MulticastAsyncCall_Passes()
         {
